+++ rest/src/main/java/com/wherecar/rest/domain/CarLog.java
@@ -24,8 +24,8 @@ public class CarLog extends BaseEntity{
     @Enumerated(EnumType.STRING)
     private GpsConditionType onGpsCondition;
 
-    private Integer onLatitude;
-    private Integer onLongitude;
+    private Double onLatitude;
+    private Double onLongitude;
     private Integer onAngle;
     private Integer onSpeed;
     private Integer onSum;
@@ -35,8 +35,8 @@ public class CarLog extends BaseEntity{
     @Enumerated(EnumType.STRING)
     private GpsConditionType offGpsCondition;
 
-    private Integer offLatitude;
-    private Integer offLongitude;
+    private Double offLatitude;
+    private Double offLongitude;
     private Integer offAngle;
     private Integer offSpeed;
     private Integer offSum;
